+++ pkgs/google_generative_ai/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.4-wip
 
-- Allow specifying an API version.
+- Allow specifying an API version in a `requestOptions` argument when
+  constructing a model.
 
 ## 0.2.3